# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vehicle Monitoring SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertType } from './alertType';

export interface Alert {
  id: number;
  vehicleId: number;
  /** @nullable */
  vehiclePlate?: string | null;
  type: AlertType;
  message: string;
  read: boolean;
  createdAt: Date;
}