# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vehicle Monitoring SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface Position {
  id: number;
  vehicleId: number;
  latitude: number;
  longitude: number;
  /** @nullable */
  speed?: number | null;
  /** @nullable */
  course?: number | null;
  /** @nullable */
  ignition?: boolean | null;
  createdAt: Date;
}