# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vehicle Monitoring SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { Position } from './position';
import type { VehicleStatus } from './vehicleStatus';

export interface Vehicle {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  plate: string;
  model: string;
  imei: string;
  status: VehicleStatus;
  lastPosition?: Position;
  createdAt: Date;
}