# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vehicle Monitoring SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleInput {
  /** @nullable */
  clientId?: number | null;
  plate: string;
  model: string;
  imei: string;
}